simple_cache_implementation/requirements.txt
redis>=4.5.0
aioredis>=2.0.0
python-dotenv>=1.0.0 